Handle null values correctly on save
This may happen, when the attribute is not contained in the list of
attributes to save (in the edit mask palette).
The MetaModel will then iterate over all attributes and save them with
the current value, which is `NULL` in this case, as no one has it
populated ever.
It would be better to initialize all attributes with an "empty" value
when creating a new item, yet this can not be done with current API.